refactor: extract in-game menu system into minarch_menu.c
Move MSG_*, Menu_init/quit/loop/beforeSleep/afterSleep,
Menu_options/scale/saveState/loadState/screenshot, achievement
comparison/display functions, getAlias, and OptionSaveChanges/
OptionAchievements helpers out of minarch.c into dedicated module.

Add Menu_setCoreVersionDesc() and Menu_waitScreenshot() wrappers to
expose static internal state (options_menu, screenshotsavethread)
to minarch.c without leaking implementation details.
