+++ workspace/all/minarch/minarch_input.h
@@ -7,9 +7,5 @@ void input_poll_callback(void);
 int16_t input_state_callback(unsigned port, unsigned device, unsigned index, unsigned id);
 void Input_init(const struct retro_input_descriptor *vars);
 
-// Menu functions defined in minarch.c, called from minarch_input.c
-void Menu_beforeSleep(void);
-void Menu_afterSleep(void);
-void Menu_screenshot(void);
-void Menu_saveState(void);
-void Menu_loadState(void);
+// Menu functions (defined in minarch_menu.c, see minarch_menu.h)
+#include "minarch_menu.h"
